pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ticktask"
version = "0.1.0"
description = "A Python CLI tool for seamless TickTick task management integration with Claude Code and Obsidian"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["ticktick", "task-management", "productivity", "obsidian", "cli", "ticktask"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dateutil",
    "pyyaml>=6.0",
    "cryptography",
    "rich>=13.0.0",
    "SQLAlchemy>=2.0",
    "aiofiles",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "types-PyYAML",
    "types-python-dateutil",
]

[project.scripts]
ticktask = "src.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/ticktask"
Documentation = "https://github.com/yourusername/ticktask#readme"
Repository = "https://github.com/yourusername/ticktask"
Issues = "https://github.com/yourusername/ticktask/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["api", "commands", "integrations", "utils"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"